+++ Sphere.hpp
@@ -6,7 +6,7 @@
 using namespace Eigen;
 
 struct Sphere{
-  Vector3d v0;
+  Vector3f v0;
   float r;
 public:
   bool isIntersect(const Ray& ray) const;
